requirements.txt
apache-airflow[gcp]
mock
